AddQ also adds when Q is empty
Previously we assumed that Q cannot be empty before AddQ is called. We simply returned false whenever this happens without adding the elem. But while swapping context, it might so happen that Q can be empty when in between the program and not just at the beginning. Hence this update is necessary.